Before changing data structure
Implemented a better data structure for the singleton use and JSON file use.

GameDataManager (singleton) should be the primary source of player data during gameplay.
The JSON file should only be used to load data at the start and save data at the end.
PlayerData should only serve as a structure for reading/writing data to/from the JSON file.